##### Changelog:
* Optimization: Improve performance for worker's initialization and memory usage.
* Feature: Clone/download project directly from source on `repo_sync_on_run` enabled.
* Feature: Set `ANSIBLE_CONFIG` env variable in project.
* Feature: Provide setup `max_content_length` in config for TAR downloading.
* Fix: Reimplement `repo_sync_on_run_timeout`.
* Fix: Bug with updating execution template from host list inventory to FK inventory.
* Chore: Bump `requests` lib version to `2.27.1`.
* Chore: Bump `vstutils` to `5.0.13`.

Closes: vstconsulting/polemarch#117
Closes: vstconsulting/polemarch#124

See merge request polemarch/ce!276

+++ polemarch/main/repo/tar.py
@@ -1,9 +1,10 @@
 # pylint: disable=expression-not-assigned,abstract-method,import-error
 from __future__ import unicode_literals
+import io
 from typing import Text, Tuple
 import tarfile
 from pathlib import Path
-from ._base import _ArchiveRepo, shutil, FILENAME
+from ._base import _ArchiveRepo, shutil
 
 
 class Tar(_ArchiveRepo):
@@ -20,7 +21,7 @@ class Tar(_ArchiveRepo):
 
         return filter(bool, map(change_member, members))
 
-    def _extract(self, archive: FILENAME, path: Text, options) -> Tuple[Path, bool]:
+    def _extract(self, archive: io.BytesIO, path: Text, options) -> Tuple[Path, bool]:
         # pylint: disable=broad-except
         moved = False
         try:
@@ -29,11 +30,12 @@ class Tar(_ArchiveRepo):
         except IOError:
             pass
         try:
-            repo_branch = self.proj.vars.get('repo_branch', '')
-            with tarfile.open(archive) as arch:
+            repo_branch = options.get('revision', self.proj.vars.get('repo_branch', ''))
+            with tarfile.open(fileobj=archive) as arch:
                 arch.extractall(path, members=self.__get_members(arch, repo_branch))
         except:
-            self.delete()
+            if not options.get('no_update', False):
+                self.delete()
             shutil.move(path + ".bak", path) if moved else None
             raise
         else:
